requirements.txt
alembic==1.0.10
Babel==2.7.0
Click==7.0
dominate==2.3.5
Flask==1.0.3
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-Table==0.5.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
nltk==3.4.3
numpy==1.16.4
pandas==0.24.2
pkg-resources==0.0.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
visitor==0.1.3
Werkzeug==0.15.4
WTForms==2.2.1
alembic==1.0.10
Babel==2.7.0
Click==7.0
dominate==2.3.5
Flask==1.0.3
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-Table==0.5.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
nltk==3.4.3
numpy==1.16.4
pandas==0.24.2
pkg-resources==0.0.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
visitor==0.1.3
Werkzeug==0.15.4
WTForms==2.2.1
